fix(task): synliggjør soft-delete-kjøring
Logger start og 0-resultat på info-nivå slik at dev kan se at SoftDeleteOppfolgingsplanerTask faktisk kjører.
